requirements.txt
slack-bolt>=1.18.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
# Для типизации в MCP клиенте
typing-extensions>=4.0.0 